Fix marketCap precision, NaN serialization and deserialization errors

+++ domain/src/main/kotlin/net/bobinski/portfolio/domain/model/Analysis.kt
@@ -16,7 +16,7 @@ data class Analysis(
     val pbRatio: Float?,
     val eps: Float?,
     val roe: Float?,
-    val marketCap: Float?
+    val marketCap: Double?
 ) {
 
     @Serializable
